+++ changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- [Add SOCKS4 and SOCKS5 proxy support for the HTTP client](https://github.com/ballerina-platform/ballerina-library/issues/8810)
+
 ### Fixed
 
 - [Fix h2c connection eviction returning upgraded channel to pool](https://github.com/ballerina-platform/ballerina-library/issues/8780)